:construction: Finished up basic modal upload feature
